Changes to have see_feedback done to access client feedback from barber side to the pages.
Now Users can send feedback from the Contact Us page and we will recieve it on the see_feedback.php page, from where barbers can answer the questions. Questions are stored in a database and get deleted when they are answered.

Answering questions sends an email to the client who asked it with the answer